[MIG] rma_sale: Migration to 17.0
TT52353
