+++ rma_sale/i18n/es.po
@@ -490,40 +490,3 @@ msgid ""
 msgstr ""
 "Va a realizar una petición de RMA. Nuestro equipo la procesará y se pondrá "
 "en contacto con usted una vez validad. Tenga en cuenta que:"
-
-#~ msgid "<strong>Requested operation:</strong>"
-#~ msgstr "<strong>Operación solicitada:</strong>"
-
-#~ msgid "<strong>Sale order:</strong>"
-#~ msgstr "<strong>Orden de venta:</strong>"
-
-#~ msgid "Stock Move"
-#~ msgstr "Movimiento de existencias"
-
-#~ msgid ""
-#~ "<span>\n"
-#~ "                                If an RMA has already been created for a "
-#~ "product in this sales order, it will not\n"
-#~ "                                be possible to create another one from "
-#~ "the web portal.\n"
-#~ "                                </span>"
-#~ msgstr ""
-#~ "<span>\n"
-#~ "                                Si ya se ha creado un RMA para un "
-#~ "producto de esta orden de venta, no será\n"
-#~ "                                posible crear otro desde el portal web.\n"
-#~ "                                </span>"
-
-#~ msgid "Domain Move"
-#~ msgstr "Dominio de movimientos"
-
-#~ msgid "Domain Picking"
-#~ msgstr "Dominio de entregas"
-
-#~ msgid "Domain Product"
-#~ msgstr "Dominio de productos"
-
-#, fuzzy
-#~| msgid "Delivery"
-#~ msgid "Is Sale Delivery"
-#~ msgstr "Es entrega de orden de venta"
