+++ rma_sale/i18n/pt_BR.po
@@ -472,24 +472,3 @@ msgid ""
 "You're about to perform an RMA request. Our team will process it an will "
 "reach you once it's validated. Keep in mind that:"
 msgstr ""
-
-#~ msgid "<strong>Requested operation:</strong>"
-#~ msgstr "<strong>Operação Requerida:</strong>"
-
-#~ msgid "<strong>Sale order:</strong>"
-#~ msgstr "<strong>Pedido de Venda:</strong>"
-
-#~ msgid ""
-#~ "<span>\n"
-#~ "                                If an RMA has already been created for a "
-#~ "product in this sales order, it will not\n"
-#~ "                                be possible to create another one from "
-#~ "the web portal.\n"
-#~ "                                </span>"
-#~ msgstr ""
-#~ "<span>\n"
-#~ "                                Se um RMA já houver sido criado para um "
-#~ "produto no pedido de venda,\n"
-#~ "                                não será possível criar outro através do "
-#~ "portal web.\n"
-#~ "                                </span>"
