+++ rma_sale/readme/USAGE.md
@@ -1,0 +1,32 @@
+To use this module, you need to:
+
+1.  Go to *RMA \> Orders* and create a new RMA.
+2.  Select a sales order to be linked to the RMA if you want.
+3.  Now you can do the rest of the instructions described in the
+    *readme* of the rma module.
+
+If you want to create one or more RMAs from a sale order:
+
+1.  Go to *Sales \> Orders \> Orders*.
+2.  Create a new sales order or select an existing one.
+3.  If the sales order is in 'Sales Order' state you can see in the
+    status bar a button labeled 'Create RMA', click it and a wizard will
+    appear.
+4.  Modify the data at your convenience and click on 'Accept' button.
+5.  As many RMAs as lines with quantity greater than zero will be
+    created. Those RMAs will be linked to the sales order.
+
+The customer can also create RMAs from a sales order portal page:
+
+1.  Go to a confirmed sales order portal page.
+2.  In the left sidebar you can see a button named 'Request RMAs'.
+3.  By clicking on this button a popup will appear to allow you to
+    define the quantity per product and delivery order line.
+4.  Click on the 'Request RMAs' button and RMAs will be created linked
+    to the sales order.
+
+Single page RMA request:
+
+1.  Go to Sales > Configuration > Settings
+2.  Check the 'Single page RMA request' box.
+3.  On a sales order in the portal, clicking on 'Request RMAs' button will open a new page (not a popup).
