+++ rma_sale/static/description/index.html
@@ -416,6 +416,13 @@ the quantity per product and delivery order line.</li>
 <li>Click on the ‘Request RMAs’ button and RMAs will be created linked to
 the sales order.</li>
 </ol>
+<p>Single page RMA request:</p>
+<ol class="arabic simple">
+<li>Go to Sales &gt; Configuration &gt; Settings</li>
+<li>Check the ‘Single page RMA request’ box.</li>
+<li>On a sales order in the portal, clicking on ‘Request RMAs’ button
+will open a new page (not a popup).</li>
+</ol>
 </div>
 <div class="section" id="known-issues-roadmap">
 <h1><a class="toc-backref" href="#toc-entry-2">Known issues / Roadmap</a></h1>
