+++ rma_sale/views/res_config_settings_views.xml
@@ -4,10 +4,7 @@
         <field name="model">res.config.settings</field>
         <field name="inherit_id" ref="sale.res_config_settings_view_form" />
         <field name="arch" type="xml">
-            <xpath
-                expr="//div[@data-key='sale_management']/div[hasclass('o_settings_container')]"
-                position="inside"
-            >
+            <block name="catalog_setting_container" position="inside">
                 <div
                     class="col-12 col-lg-6 o_setting_box"
                     title="Show portal RMA request in a single page"
@@ -30,7 +27,7 @@
                         </div>
                     </div>
                 </div>
-            </xpath>
+            </block>
         </field>
     </record>
 </odoo>
