+++ rma_sale/views/sale_views.xml
@@ -10,7 +10,7 @@
                     name="action_create_rma"
                     type="object"
                     string="Create RMA"
-                    states="sale,done"
+                    invisible="state not in ('sale','done')"
                     groups="rma.rma_group_user_own"
                 />
             </button>
@@ -20,7 +20,7 @@
                     type="object"
                     class="oe_stat_button"
                     icon="fa-reply"
-                    attrs="{'invisible': [('rma_count', '=', 0)]}"
+                    invisible="rma_count==0"
                     groups="rma.rma_group_user_own"
                 >
                     <field name="rma_count" widget="statinfo" string="RMA" />
