+++ rma_sale/wizard/sale_order_rma_wizard_views.xml
@@ -31,10 +31,7 @@
                                 options="{'no_create': True}"
                                 required="1"
                             />
-                            <field
-                                name="operation_id"
-                                attrs="{'required': [('quantity', '>', 0)]}"
-                            />
+                            <field name="operation_id" required="quantity>0" />
                         </tree>
                     </field>
                 </group>
